Update to last files formats (signals, models, admin, apps, worker)
