+++ pyscada/gpio/apps.py
@@ -10,3 +10,6 @@ class PyScadaGPIOConfig(AppConfig):
     name = 'pyscada.gpio'
     verbose_name = _("PyScada GPIO")
     path = os.path.dirname(os.path.realpath(__file__))
+
+    def ready(self):
+        import pyscada.gpio.signals
